Revert "Refactor FrameSeq. Fix MIDI sync generation."
