+++ src/main/lcdgui/screens/UserScreen.hpp
@@ -14,7 +14,7 @@ namespace mpc::sequencer
 {
 	class Sequence;
 	class Sequencer;
-	class FrameSeqBase;
+	class FrameSeq;
 }
 
 namespace mpc::nvram
@@ -105,7 +105,7 @@ namespace mpc::lcdgui::screens
 		friend class mpc::nvram::NvRam;
 		friend class mpc::sequencer::Sequence;
 		friend class mpc::sequencer::Sequencer;
-		friend class mpc::sequencer::FrameSeqBase;
+		friend class mpc::sequencer::FrameSeq;
 		friend class mpc::Util;
 
 	};
