+++ src/main/sequencer/Clock.hpp
@@ -2,35 +2,35 @@
 
 #include <cmath>
 
-namespace mpc::sequencer {
+namespace mpc {
+	namespace sequencer {
 
-    class Clock
-    {
+		class Clock {
 
-    private:
-        static const double ppqn_f;
-        static const double rec60;
-        static const double rec03;
-        double bpm, omega, pd;
-        double vpd, lvpd, dinphase;
-        int tickN; // the tick number
-        double Fs_rec;
+		private:
+			static const double ppqn_f;
+			static const double rec60;
+			static const double rec03;
+			double bpm, omega, pd, lpd;
+			double vpd, lvpd, dinphase;
+			int tickN; // the tick number
+			double Fs_rec;
+			int cycleCount;
+			int tickOffset{ 0 };
 
-    public:
-        void init(double Fs);
+		public:
+			void init(double Fs);
+			void set_bpm(const double &BPM);
+            double getSampleRate() { return 1.0 / Fs_rec; }
+			bool proc();
+			void reset();
+			void zero();
+			
+			int getTickN();
+			void setTick(int i);
+			int getTickPosition();
+			double getBpm();
+		};
 
-        void set_bpm(const double &BPM);
-
-        double getSampleRate() const
-        { return 1.0 / Fs_rec; }
-
-        bool proc();
-
-        void reset();
-
-        void zero();
-
-        double getBpm() const;
-    };
-
+	}
 }
