+++ src/main/sequencer/Event.cpp
@@ -1,6 +1,13 @@
 #include <sequencer/Event.hpp>
+#include <midi/core/MidiMessage.hpp>
+#include <midi/core/ShortMessage.hpp>
 
 using namespace mpc::sequencer;
+
+Event::Event()
+{
+	shortMessage = new ctoot::midi::core::ShortMessage();
+}
 
 void Event::setTick(int relativeTick)
 {
@@ -28,4 +35,14 @@ void Event::CopyValuesTo(std::weak_ptr<Event> dest) {
 	auto lDest = dest.lock();
 	lDest->setTick(getTick());
 	lDest->setTrack(getTrack());
+}
+
+ctoot::midi::core::ShortMessage* Event::getShortMessage()
+{
+	return shortMessage;
+}
+
+
+Event::~Event() {
+	delete shortMessage;
 }
